fixed state save weird char

+++ GameObject.cpp
@@ -63,7 +63,7 @@ void GameObject::save(ofstream &file) const
     file << display_code << " ";
     file << id_num << " ";
     file << location.x << " " << location.y << " ";
-    file << state << " ";
+    file << (int)state << " ";
 }
 
 void GameObject::restore(ifstream &file, Model &model) const
